Fix RHEL 6 compiler issue.

+++ mgmt/api/EventControlMain.cc
@@ -550,7 +550,7 @@ handle_event_message(EventClientT * client, void * req, size_t reqlen)
 {
   OpType optype = extract_mgmt_request_optype(req, reqlen);
 
-  if (optype < 0 || optype >= countof(handlers)) {
+  if (optype < 0 || static_cast<unsigned>(optype) >= countof(handlers)) {
     goto fail;
   }
